Move route components to dedicated "routes" folder
