+++ frontend/src/features/auth/components/routes/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from "react";
 import { Navigate } from "react-router-dom";
-import useIsAuthenticated from "../hooks/useIsAuthenticated";
+import useIsAuthenticated from "../../hooks/useIsAuthenticated";
 
 interface ProtectedRouteProps {
 	children: ReactNode;
@@ -10,7 +10,7 @@ function ProtectedRoute({ children }: ProtectedRouteProps) {
 	const isAuthenticated = useIsAuthenticated();
 
 	if (isAuthenticated === null) {
-		return <div>Loading</div>
+		return <div>Loading</div>;
 	}
 
 	return isAuthenticated ? children : <Navigate to={"/login"} />;
